[main] train and evaluate model

+++ BE/src/main/java/com/flowmind/domain/dataset/dto/AnnotationDto.java
@@ -1,9 +1,7 @@
 package com.flowmind.domain.dataset.dto;
 
-import java.util.List;
-
 public record AnnotationDto(
-        Long id,
+        Integer id,
         String label,
         double xCenter,
         double yCenter,
