+++ BE/src/main/java/com/flowmind/domain/dataset/dto/ImageWithAnnotationsDto.java
@@ -3,8 +3,8 @@ package com.flowmind.domain.dataset.dto;
 import java.util.List;
 
 public record ImageWithAnnotationsDto(
-        Long id,
+        Integer id,
         String filename,
-        String imageUrl,              // 실제 이미지 요청용 URL
+        String imageUrl,
         List<AnnotationDto> annotations
 ) {}
