+++ BE/src/main/java/com/flowmind/domain/user/repository/UserRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.flowmind.domain.user.entity.User;
 
-public interface UserRepository extends JpaRepository<User, Long>{
-	boolean existsByEmail(String email);
-	boolean existsByPhone(String phone);
-	
-	Optional<User> findByEmail(String email);
+public interface UserRepository extends JpaRepository<User, Integer> {
+    boolean existsByEmail(String email);
+    boolean existsByPhone(String phone);
+
+    Optional<User> findByEmail(String email);
 }
