+++ BE/src/main/java/com/flowmind/security/JwtAuthFilter.java
@@ -30,33 +30,24 @@ public class JwtAuthFilter extends OncePerRequestFilter {
 
         final String authHeader = request.getHeader("Authorization");
 
-        // Token 없으면 다음 필터로 넘김
         if (authHeader == null || !authHeader.startsWith("Bearer ")) {
             filterChain.doFilter(request, response);
             return;
         }
 
-        String token = authHeader.substring(7); // "Bearer " 이후
+        String token = authHeader.substring(7);
 
         try {
             Claims claims = jwtUtil.getClaims(token);
-
-            Long userId = claims.get("id", Long.class);
+            Integer userId = claims.get("id", Integer.class);
             String email = claims.get("email", String.class);
 
             UsernamePasswordAuthenticationToken authentication =
-                    new UsernamePasswordAuthenticationToken(
-                            email, null, null
-                    );
+                    new UsernamePasswordAuthenticationToken(email, null, null);
 
-            authentication.setDetails(
-                    new WebAuthenticationDetailsSource().buildDetails(request)
-            );
-
+            authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
             SecurityContextHolder.getContext().setAuthentication(authentication);
-
         } catch (Exception e) {
-            // 토큰이 잘못되었으면 SecurityContext 초기화
             SecurityContextHolder.clearContext();
         }
